Cleanup demo_bevy
