+++ demo_bevy/src/lib.rs
@@ -8,6 +8,10 @@ use serde::{Deserialize, Serialize};
 pub const PRIVATE_KEY: &[u8; NETCODE_KEY_BYTES] = b"an example very very secret key."; // 32-bytes
 pub const PROTOCOL_ID: u64 = 7;
 
+pub mod client;
+
+pub mod server;
+
 #[derive(Debug, Component)]
 pub struct Player {
     pub id: u64,
